+++ lib/services/cache-warming-rule-service.ts
@@ -1,0 +1,458 @@
+import { UnifiedCacheManager } from "./cache-orchestrator";
+import { CacheKeyGeneratorService } from "./cache-key-generator-service";
+import { logger } from "@/lib/logger";
+import type { AIPatternType, CacheWarmingRule, AIPattern } from "./ai-pattern-types";
+
+/**
+ * Cache Warming Rule Service
+ * Manages cache warming rules and execution
+ */
+export class CacheWarmingRuleService {
+  private static readonly WARMING_RULES: CacheWarmingRule[] = [
+    {
+      pattern: "marketplace",
+      triggers: ["marketplace", "seller platform", "multi-vendor"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL",
+          auth: "Clerk",
+          payments: "Stripe",
+          deployment: "Vercel",
+        },
+        features: [
+          "Multi-vendor product listings",
+          "User authentication and profiles",
+          "Search and filtering",
+          "Commission management",
+          "Review and rating system",
+          "Payment processing",
+          "Admin dashboard",
+        ],
+        architecture: {
+          type: "Microservices",
+          scaling: "Serverless-ready",
+          database: "PostgreSQL with Redis caching",
+          cdn: "Vercel Edge Network",
+        },
+        monetization: [
+          { type: "Commission", rate: "5-15%" },
+          { type: "Listings", price: "$29-99/month" },
+          { type: "Transaction", rate: "2-5%" },
+        ],
+        estimatedLines: 25000,
+        complexity: "high",
+      },
+      ttl: 7200,
+      priority: 1,
+    },
+    {
+      pattern: "ecommerce",
+      triggers: ["online store", "shopping", "ecommerce"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL",
+          auth: "Clerk",
+          payments: "Stripe",
+          deployment: "Vercel",
+        },
+        features: [
+          "Product catalog",
+          "Shopping cart",
+          "Secure checkout",
+          "Order management",
+          "Inventory tracking",
+          "Customer accounts",
+          "Payment integration",
+        ],
+        architecture: {
+          type: "Monolithic with microservices potential",
+          scaling: "Horizontal with connection pooling",
+          database: "PostgreSQL with row-level security",
+          cdn: "Vercel Edge Network",
+        },
+        monetization: [
+          { type: "Product sales", margin: "30-60%" },
+          { type: "Subscription", price: "$49-199/month" },
+        ],
+        estimatedLines: 20000,
+        complexity: "medium",
+      },
+      ttl: 3600,
+      priority: 2,
+    },
+    {
+      pattern: "social",
+      triggers: ["social network", "community platform", "social app"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL",
+          auth: "Clerk",
+          realTime: "WebSockets",
+          deployment: "Vercel",
+        },
+        features: [
+          "User profiles",
+          "Social feed",
+          "Messaging system",
+          "Follow system",
+          "Post creation",
+          "Comments and likes",
+          "Notifications",
+        ],
+        architecture: {
+          type: "Microservices",
+          scaling: "Horizontal with Redis",
+          database: "PostgreSQL with RLS",
+          cdn: "Vercel Edge Network",
+        },
+        monetization: [
+          { type: "Premium features", price: "$9-29/month" },
+          { type: "Advertising", model: "CPM/CPC" },
+        ],
+        estimatedLines: 23000,
+        complexity: "high",
+      },
+      ttl: 5400,
+      priority: 3,
+    },
+    {
+      pattern: "fintech",
+      triggers: ["banking app", "payment platform", "investment platform"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL with encryption",
+          auth: "Clerk with MFA",
+          payments: ["Stripe", "Plaid API"],
+          blockchain: ["Web3.js", "Ethereum"],
+          deployment: "AWS with SOC 2 compliance",
+        },
+        features: [
+          "Secure user authentication with MFA",
+          "Bank account integration via Plaid",
+          "Real-time payment processing",
+          "Portfolio management",
+          "Transaction history and analytics",
+          "Regulatory compliance (KYC/AML)",
+          "Multi-currency support",
+          "Audit logging and reporting",
+        ],
+        architecture: {
+          type: "Microservices with event sourcing",
+          scaling: "Horizontal with circuit breakers",
+          database: "PostgreSQL with encryption at rest",
+          compliance: ["SOC 2", "PCI DSS", "GDPR"],
+          monitoring: ["Real-time fraud detection", "Compliance monitoring"],
+        },
+        monetization: [
+          { type: "Transaction fees", rate: "0.5-2.5%" },
+          { type: "Premium features", price: "$29-299/month" },
+          { type: "API usage", model: "Pay-per-call" },
+        ],
+        estimatedLines: 35000,
+        complexity: "very-high",
+      },
+      ttl: 10800,
+      priority: 1,
+    },
+    {
+      pattern: "healthcare",
+      triggers: [
+        "telemedicine platform",
+        "healthcare app",
+        "patient management",
+      ],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL with HIPAA compliance",
+          auth: "Clerk with healthcare compliance",
+          payments: "Stripe with healthcare processing",
+          video: ["WebRTC", "Twilio Video"],
+          deployment: "AWS HIPAA-compliant",
+        },
+        features: [
+          "HIPAA-compliant patient records",
+          "Telemedicine video consultations",
+          "Appointment scheduling system",
+          "Prescription management",
+          "Insurance verification",
+          "Secure messaging with providers",
+          "Health data visualization",
+          "Emergency contact management",
+        ],
+        architecture: {
+          type: "Microservices with audit trails",
+          scaling: "High availability with failover",
+          database: "PostgreSQL with field-level encryption",
+          compliance: ["HIPAA", "HITECH", "GDPR"],
+          monitoring: ["Audit logging", "Access monitoring"],
+        },
+        monetization: [
+          { type: "Subscription", price: "$99-999/month" },
+          { type: "Per-consultation", fee: "$50-200" },
+          { type: "Enterprise licensing", custom: true },
+        ],
+        estimatedLines: 40000,
+        complexity: "very-high",
+      },
+      ttl: 7200,
+      priority: 1,
+    },
+    {
+      pattern: "edtech",
+      triggers: ["learning platform", "online education", "training system"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL",
+          auth: "Clerk with SSO support",
+          video: ["Mux", "Vimeo API"],
+          content: ["Markdown", "PDF generation"],
+          deployment: "Vercel Edge",
+        },
+        features: [
+          "Student enrollment and management",
+          "Course creation and delivery",
+          "Video lecture hosting",
+          "Interactive quizzes and assessments",
+          "Progress tracking and analytics",
+          "Certificate generation",
+          "Discussion forums",
+          "Assignment submission system",
+        ],
+        architecture: {
+          type: "Microservices",
+          scaling: "Auto-scaling for course demand",
+          database: "PostgreSQL with content caching",
+          cdn: "Global video distribution",
+          monitoring: ["Learning analytics", "Engagement tracking"],
+        },
+        monetization: [
+          { type: "Per-course", price: "$49-299" },
+          { type: "Subscription", price: "$29-99/month" },
+          { type: "Enterprise plans", custom: true },
+        ],
+        estimatedLines: 28000,
+        complexity: "high",
+      },
+      ttl: 5400,
+      priority: 2,
+    },
+    {
+      pattern: "realestate",
+      triggers: ["property management", "real estate platform", "rental app"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL with GIS support",
+          auth: "Clerk",
+          payments: "Stripe with escrow",
+          maps: ["Google Maps API", "Mapbox"],
+          deployment: "Vercel Edge",
+        },
+        features: [
+          "Property search and filtering",
+          "Interactive map listings",
+          "Virtual tour integration",
+          "Rental application system",
+          "Document management",
+          "Tenant screening",
+          "Maintenance request tracking",
+          "Property analytics dashboard",
+        ],
+        architecture: {
+          type: "Microservices",
+          scaling: "Geographically distributed",
+          database: "PostgreSQL with geospatial queries",
+          cdn: "Optimized for images and virtual tours",
+          monitoring: ["Lead tracking", "Conversion analytics"],
+        },
+        monetization: [
+          { type: "Listing fees", price: "$99-499/listing" },
+          { type: "Subscription", price: "$199-999/month" },
+          { type: "Transaction fees", rate: "1-3%" },
+        ],
+        estimatedLines: 25000,
+        complexity: "medium",
+      },
+      ttl: 7200,
+      priority: 2,
+    },
+    {
+      pattern: "logistics",
+      triggers: ["fleet management", "supply chain platform", "delivery app"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL with temporal tables",
+          auth: "Clerk",
+          maps: ["Google Maps API", "Routing APIs"],
+          realTime: ["WebSockets", "GPS tracking"],
+          deployment: "AWS with auto-scaling",
+        },
+        features: [
+          "Real-time fleet tracking",
+          "Route optimization algorithms",
+          "Inventory management system",
+          "Warehouse management",
+          "Delivery scheduling",
+          "Driver management app",
+          "Analytics and reporting",
+          "Customer notification system",
+        ],
+        architecture: {
+          type: "Microservices with event sourcing",
+          scaling: "High availability for real-time tracking",
+          database: "PostgreSQL with temporal data",
+          cdn: "Global mapping services",
+          monitoring: ["Real-time tracking", "Performance analytics"],
+        },
+        monetization: [
+          { type: "Per-vehicle", price: "$29-99/month" },
+          { type: "Per-delivery", rate: "$0.10-2" },
+          { type: "Enterprise", custom: true },
+        ],
+        estimatedLines: 32000,
+        complexity: "high",
+      },
+      ttl: 5400,
+      priority: 2,
+    },
+    {
+      pattern: "saas",
+      triggers: ["B2B software", "enterprise platform", "business software"],
+      prewarmedData: {
+        techStack: {
+          frontend: ["Next.js", "React", "TypeScript", "Tailwind CSS"],
+          backend: ["Node.js", "Express", "PostgreSQL", "Redis"],
+          database: "Neon PostgreSQL with multi-tenancy",
+          auth: "Clerk with SSO/SAML",
+          payments: "Stripe with enterprise billing",
+          analytics: ["Segment", "Mixpanel"],
+          deployment: "AWS with multi-region",
+        },
+        features: [
+          "Multi-tenant architecture",
+          "Role-based access control (RBAC)",
+          "White-label customization",
+          "Advanced analytics dashboard",
+          "API management and documentation",
+          "Enterprise SSO integration",
+          "Custom workflow builder",
+          "Advanced reporting and exports",
+        ],
+        architecture: {
+          type: "Microservices with data isolation",
+          scaling: "Horizontal with per-tenant scaling",
+          database: "PostgreSQL with RLS for multi-tenancy",
+          security: ["SOC 2", "ISO 27001", "GDPR"],
+          monitoring: ["Per-tenant monitoring", "SLA tracking"],
+        },
+        monetization: [
+          { type: "Per-seat", price: "$29-499/month" },
+          { type: "Usage-based", tier: "By API calls/users" },
+          { type: "Enterprise", custom: true },
+        ],
+        estimatedLines: 38000,
+        complexity: "very-high",
+      },
+      ttl: 3600,
+      priority: 3,
+    },
+  ];
+
+  /**
+   * Determine if a warming rule should be executed
+   */
+  static shouldWarmRule(
+    rule: CacheWarmingRule,
+    detectedPatterns: AIPattern["type"][],
+  ): boolean {
+    if (rule.priority <= 2) {
+      return true;
+    }
+
+    return detectedPatterns.includes(rule.pattern);
+  }
+
+  /**
+   * Warm a specific cache rule
+   */
+  static async warmRule(rule: CacheWarmingRule): Promise<void> {
+    try {
+      const cacheKey = CacheKeyGeneratorService.generateOptimizedCacheKey(
+        "iflow",
+        `blueprint-${rule.pattern}`,
+        rule.pattern,
+        rule.pattern.includes("-") ? rule.pattern.split("-")[0] : undefined,
+      );
+
+      await UnifiedCacheManager.setData(cacheKey, rule.prewarmedData, {
+        ttl: rule.ttl,
+        tags: ["ai-warmed", rule.pattern, "blueprint-skeleton"],
+      });
+
+      logger.debug("AI cache rule warmed", {
+        pattern: rule.pattern,
+        cacheKey,
+        ttl: rule.ttl,
+        priority: rule.priority,
+      });
+    } catch (error) {
+      logger.debug("Failed to warm AI cache rule", {
+        pattern: rule.pattern,
+        error: error instanceof Error ? error.message : "Unknown error",
+      });
+    }
+  }
+
+  /**
+   * Calculate estimated cost savings for a warming rule
+   */
+  static calculateEstimatedSavings(rule: CacheWarmingRule): number {
+    const averageRequestsPerHour = 10;
+    const hoursInTTL = rule.ttl / 3600;
+    const iflowCostPerRequest = 0.02;
+
+    return averageRequestsPerHour * hoursInTTL * iflowCostPerRequest;
+  }
+
+  /**
+   * Get all warming rules
+   */
+  static getWarmingRules(): CacheWarmingRule[] {
+    return [...CacheWarmingRuleService.WARMING_RULES];
+  }
+
+  /**
+   * Get warming rules by priority
+   */
+  static getWarmingRulesByPriority(priority?: number): CacheWarmingRule[] {
+    const rules = [...CacheWarmingRuleService.WARMING_RULES];
+    
+    if (priority !== undefined) {
+      return rules.filter((rule) => rule.priority === priority);
+    }
+    
+    return rules.sort((a, b) => a.priority - b.priority);
+  }
+
+  /**
+   * Get warming rules by pattern type
+   */
+  static getWarmingRuleByPattern(pattern: AIPatternType): CacheWarmingRule | undefined {
+    return CacheWarmingRuleService.WARMING_RULES.find((rule) => rule.pattern === pattern);
+  }
+}
